+++ miso-web/src/it/java/uk/ac/bbsrc/tgac/miso/webapp/integrationtest/ListTablesIT.java
@@ -15,7 +15,6 @@ import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import org.apache.commons.lang3.math.NumberUtils;
-import org.junit.Before;
 import org.junit.Test;
 
 import com.google.common.collect.Sets;
@@ -52,6 +51,7 @@ public class ListTablesIT extends AbstractIT {
     tabs.put(ListTarget.STORAGE_LOCATIONS, Sets.newHashSet(Tabs.FREEZERS, Tabs.ROOMS));
     tabs.put(ListTarget.POOL_ORDERS, Sets.newHashSet(Tabs.OUTSTANDING, Tabs.FULFILLED, Tabs.DRAFT));
     tabs.put(ListTarget.TRANSFERS, Sets.newHashSet(Tabs.PENDING, Tabs.RECEIPT, Tabs.INTERNAL, Tabs.DISTRIBUTION));
+    tabs.put(ListTarget.INSTRUMENTS, Sets.newHashSet(Tabs.SEQUENCER, Tabs.ARRAY_SCANNER, Tabs.OTHER));
     tabsForTarget = Collections.unmodifiableMap(tabs);
   }
 
@@ -73,6 +73,7 @@ public class ListTablesIT extends AbstractIT {
     preferredTab.put(ListTarget.WORKSETS, Tabs.MINE);
     preferredTab.put(ListTarget.POOL_ORDERS, Tabs.OUTSTANDING);
     preferredTab.put(ListTarget.TRANSFERS, Tabs.RECEIPT);
+    preferredTab.put(ListTarget.INSTRUMENTS, Tabs.SEQUENCER);
     sortOnTab = Collections.unmodifiableMap(preferredTab);
   }
 
@@ -144,11 +145,6 @@ public class ListTablesIT extends AbstractIT {
     return m1.group(1);
   }
 
-  @Before
-  public void setup() {
-    login();
-  }
-
   private ListPage getList(String listTarget) {
     return ListPage.getListPage(getDriver(), getBaseUrl(), listTarget);
   }
@@ -160,6 +156,7 @@ public class ListTablesIT extends AbstractIT {
   @Test
   public void testIndexDistanceToolSetup() throws Exception {
     // Goal: ensure all expected columns are present and no extra
+    login();
     Set<String> indicesColumns = Sets.newHashSet(Columns.FAMILY, Columns.INDEX_NAME, Columns.SEQUENCE, Columns.POSITION);
     ListPage page = ListPage.getListPage(getDriver(), getBaseUrl(), "tools/indexdistance");
     DataTable table = page.getTable();
@@ -176,6 +173,7 @@ public class ListTablesIT extends AbstractIT {
   @Test
   public void testIdentitySearchToolSetup() throws Exception {
     // Goal: ensure all expected columns are present and no extra
+    login();
     ListPage page = ListPage.getListPage(getDriver(), getBaseUrl(), "tools/identitysearch");
     DataTable table = page.getTable();
     List<String> headings = table.getColumnHeadings();
@@ -213,6 +211,7 @@ public class ListTablesIT extends AbstractIT {
 
   @Test
   public void testListLibrariesWarnings() throws Exception {
+    login();
     testWarningNormal(ListTarget.LIBRARIES, "LIB901", "Negative Volume", Columns.NAME);
   }
 
@@ -246,6 +245,7 @@ public class ListTablesIT extends AbstractIT {
 
   @Test
   public void testListPoolsWarnings() throws Exception {
+    login();
     testWarningTabbed(ListTarget.POOLS, "no indices", "MISSING INDEX", Columns.NAME);
     testWarningTabbed(ListTarget.POOLS, "similar index", "Near-Duplicate Indices", Columns.NAME);
     testWarningTabbed(ListTarget.POOLS, "same index", "DUPLICATE INDICES", Columns.NAME);
@@ -254,6 +254,7 @@ public class ListTablesIT extends AbstractIT {
 
   @Test
   public void testListOrdersSetup() throws Exception {
+    login();
     Set<String> ordersColumns = Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.ALIAS, Columns.PURPOSE, Columns.ORDER_DESCRIPTION,
         Columns.POOL_DESCRIPTION, Columns.INSTRUMENT_MODEL, Columns.LONGEST_INDEX, Columns.SEQUENCING_PARAMETERS, Columns.REMAINING,
         Columns.LAST_MODIFIED);
@@ -281,13 +282,15 @@ public class ListTablesIT extends AbstractIT {
 
   @Test
   public void testListOrdersColumnSort() throws Exception {
-    testTabbedColumnsSort(ListTarget.ORDERS_ALL);
-    testTabbedColumnsSort(ListTarget.ORDERS_OUTSTANDING);
-    testTabbedColumnsSort(ListTarget.ORDERS_IN_PROGRESS);
+    login();
+    testTabbedColumnsSort(ListTarget.ORDERS_ALL, true);
+    testTabbedColumnsSort(ListTarget.ORDERS_OUTSTANDING, true);
+    testTabbedColumnsSort(ListTarget.ORDERS_IN_PROGRESS, true);
   }
 
   @Test
   public void testListOrdersWarnings() throws Exception {
+    login();
     testWarningTabbed(ListTarget.ORDERS_ALL, "no indices", "MISSING INDEX", Columns.POOL_DESCRIPTION);
     testWarningTabbed(ListTarget.ORDERS_ALL, "similar index", "Near-Duplicate Indices", Columns.POOL_DESCRIPTION);
     testWarningTabbed(ListTarget.ORDERS_ALL, "same index", "DUPLICATE INDICES", Columns.POOL_DESCRIPTION);
@@ -329,13 +332,13 @@ public class ListTablesIT extends AbstractIT {
 
   @Test
   public void testListInstrumentsSetup() throws Exception {
-    testPageSetup(ListTarget.INSTRUMENTS,
+    testTabbedPageSetup(ListTarget.INSTRUMENTS,
         Sets.newHashSet(Columns.INSTRUMENT_NAME, Columns.PLATFORM, Columns.INSTRUMENT_MODEL, Columns.STATUS, Columns.SERIAL_NUMBER));
   }
 
   @Test
   public void testListSequencersColumnSort() throws Exception {
-    testColumnsSort(ListTarget.INSTRUMENTS);
+    testTabbedColumnsSort(ListTarget.INSTRUMENTS);
   }
 
   @Test
@@ -609,7 +612,7 @@ public class ListTablesIT extends AbstractIT {
 
   @Test
   public void testListTabbedStorageLocationsSetup() throws Exception {
-    testTabbedPageSetup(ListTarget.STORAGE_LOCATIONS, Sets.newHashSet(Columns.FREEZER_NAME, Columns.IDENTIFICATION_BARCODE, Columns.MAP));
+    testTabbedPageSetup(ListTarget.STORAGE_LOCATIONS, Sets.newHashSet(Columns.ALIAS, Columns.IDENTIFICATION_BARCODE, Columns.MAP));
   }
 
   @Test
@@ -700,8 +703,166 @@ public class ListTablesIT extends AbstractIT {
     testTabbedColumnsSort(ListTarget.TRANSFERS);
   }
 
+  @Test
+  public void testListLibraryTemplatesSetup() throws Exception {
+    testPageSetup(ListTarget.LIBRARY_TEMPLATES,
+        Sets.newHashSet(Columns.SORT, Columns.ALIAS, Columns.LIBRARY_DESIGN, Columns.LIBRARY_DESIGN_CODE, Columns.LIBRARY_TYPE,
+            Columns.LIBRARY_SELECTION, Columns.LIBRARY_STRATEGY, Columns.KIT_NAME, Columns.INDEX_FAMILY, Columns.PLATFORM,
+            Columns.DEFAULT_VOLUME));
+  }
+
+  @Test
+  public void testListLibraryTemplatesColumnSort() throws Exception {
+    testColumnsSort(ListTarget.LIBRARY_TEMPLATES);
+  }
+
+  @Test
+  public void testListQcTypesSetup() throws Exception {
+    testPageSetup(ListTarget.QC_TYPE, Sets.newHashSet(Columns.NAME, Columns.DESCRIPTION, Columns.TARGET, Columns.UNITS,
+        Columns.CORRESPONDING_FIELD, Columns.AUTO_UPDATE_FIELD));
+  }
+
+  @Test
+  public void testListQcTypesColumnSort() throws Exception {
+    testColumnsSort(ListTarget.QC_TYPE);
+  }
+
+  @Test
+  public void testListAttachmentCategoriesSetup() throws Exception {
+    testPageSetup(ListTarget.ATTACHMENT_CATEGORIES, Sets.newHashSet(Columns.ALIAS));
+  }
+
+  @Test
+  public void testListAttachmentCategoriesColumnSort() throws Exception {
+    testColumnsSort(ListTarget.ATTACHMENT_CATEGORIES);
+  }
+
+  @Test
+  public void testListInstitutesSetup() throws Exception {
+    testPageSetup(ListTarget.INSTITUTES, Sets.newHashSet(Columns.ALIAS));
+  }
+
+  @Test
+  public void testListInstitutesColumnSort() throws Exception {
+    testColumnsSort(ListTarget.INSTITUTES);
+  }
+
+  @Test
+  public void testListLabsSetup() throws Exception {
+    testPageSetup(ListTarget.LABS, Sets.newHashSet(Columns.ALIAS, Columns.INSTITUTE));
+  }
+
+  @Test
+  public void testListLabsColumnSort() throws Exception {
+    testColumnsSort(ListTarget.LABS);
+  }
+
+  @Test
+  public void testListTissueMaterialsSetup() throws Exception {
+    testPageSetup(ListTarget.TISSUE_MATERIALS, Sets.newHashSet(Columns.ALIAS));
+  }
+
+  @Test
+  public void testListTissueMaterialsColumnSort() throws Exception {
+    testColumnsSort(ListTarget.TISSUE_MATERIALS);
+  }
+
+  @Test
+  public void testListTissueOriginsSetup() throws Exception {
+    testPageSetup(ListTarget.TISSUE_ORIGINS, Sets.newHashSet(Columns.ALIAS, Columns.DESCRIPTION));
+  }
+
+  @Test
+  public void testListTissueOriginsColumnSort() throws Exception {
+    testColumnsSort(ListTarget.TISSUE_ORIGINS);
+  }
+
+  @Test
+  public void testListTissuePieceTypesSetup() throws Exception {
+    testPageSetup(ListTarget.TISSUE_PIECE_TYPE, Sets.newHashSet(Columns.NAME, Columns.ABBREVIATION, Columns.ARCHIVED));
+  }
+
+  @Test
+  public void testListTissuePieceTypesColumnSort() throws Exception {
+    testColumnsSort(ListTarget.TISSUE_PIECE_TYPE);
+  }
+
+  @Test
+  public void testListSamplePurposesSetup() throws Exception {
+    testPageSetup(ListTarget.SAMPLE_PURPOSES, Sets.newHashSet(Columns.ALIAS));
+  }
+
+  @Test
+  public void testListSamplePurposesColumnSort() throws Exception {
+    testColumnsSort(ListTarget.SAMPLE_PURPOSES);
+  }
+
+  @Test
+  public void testListSubprojectsSetup() throws Exception {
+    testPageSetup(ListTarget.SUBPROJECTS, Sets.newHashSet(Columns.ALIAS, Columns.PROJECT, Columns.PRIORITY, Columns.REFERENCE_GENOME));
+  }
+
+  @Test
+  public void testListSubprojectsColumnSort() throws Exception {
+    testColumnsSort(ListTarget.SUBPROJECTS);
+  }
+
+  @Test
+  public void testListExperimentsSetup() throws Exception {
+    testPageSetup(ListTarget.EXPERIMENTS, Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.ALIAS, Columns.PLATFORM, Columns.LIBRARY_NAME,
+        Columns.LIBRARY_ALIAS, Columns.STUDY_NAME, Columns.STUDY_ALIAS));
+  }
+
+  @Test
+  public void testListExperimentsColumnSort() throws Exception {
+    testColumnsSort(ListTarget.EXPERIMENTS);
+  }
+
+  @Test
+  public void testListSubmissionsSetup() throws Exception {
+    testPageSetup(ListTarget.SUBMISSIONS, Sets.newHashSet(Columns.ID, Columns.ALIAS, Columns.CREATION_DATE, Columns.SUBMISSION_DATE,
+        Columns.VERIFIED, Columns.COMPLETED));
+  }
+
+  @Test
+  public void testListSubmissionsColumnSort() throws Exception {
+    testColumnsSort(ListTarget.SUBMISSIONS);
+  }
+
+  @Test
+  public void testListUsersSetup() throws Exception {
+    loginAdmin();
+    testPageSetup(ListTarget.USERS, Sets.newHashSet(Columns.SORT, Columns.LOGIN_NAME, Columns.FULL_NAME, Columns.ACTIVE, Columns.ADMIN,
+        Columns.INTERNAL, Columns.LOGGED_IN), true);
+  }
+
+  @Test
+  public void testListUsersColumnSort() throws Exception {
+    loginAdmin();
+    testColumnsSort(ListTarget.USERS, true);
+  }
+
+  @Test
+  public void testListGroupsSetup() throws Exception {
+    loginAdmin();
+    testPageSetup(ListTarget.GROUPS, Sets.newHashSet(Columns.SORT, Columns.NAME, Columns.DESCRIPTION), true);
+  }
+
+  @Test
+  public void testListGroupsColumnSort() throws Exception {
+    loginAdmin();
+    testColumnsSort(ListTarget.GROUPS, true);
+  }
+
   private void testPageSetup(String listTarget, Set<String> targetColumns) {
+    testPageSetup(listTarget, targetColumns, false);
+  }
+
+  private void testPageSetup(String listTarget, Set<String> targetColumns, boolean skipLogin) {
     // Goal: confirm that all expected columns are present
+    if (!skipLogin) {
+      login();
+    }
     ListPage page = getList(listTarget);
     DataTable table = page.getTable();
     List<String> headings = table.getColumnHeadings();
@@ -713,6 +874,7 @@ public class ListTablesIT extends AbstractIT {
 
   private void testTabbedPageSetup(String listTarget, Set<String> targetColumns) {
     // Goal: confirm that all expected tabs and columns are present
+    login();
     ListTabbedPage page = getTabbedList(listTarget);
     DataTable table = page.getTable();
     // confirm expected number of tabs
@@ -730,14 +892,28 @@ public class ListTablesIT extends AbstractIT {
   }
 
   private void testColumnsSort(String listTarget) {
+    testColumnsSort(listTarget, false);
+  }
+
+  private void testColumnsSort(String listTarget, boolean skipLogin) {
     // confirm that sortable columns can be sorted on
+    if (!skipLogin) {
+      login();
+    }
     ListPage page = getList(listTarget);
     sortColumns(page.getTable(), page);
   }
 
   private void testTabbedColumnsSort(String listTarget) {
+    testTabbedColumnsSort(listTarget, false);
+  }
+
+  private void testTabbedColumnsSort(String listTarget, boolean skipLogin) {
     // confirm that sortable columns can be sorted on
     // note that this sorts in a single tab only, as different tabs should not have different columns.
+    if (!skipLogin) {
+      login();
+    }
     ListTabbedPage page = getTabbedList(listTarget);
     DataTable table = page.getTable();
     sortColumns(table, page);
@@ -820,6 +996,9 @@ public class ListTablesIT extends AbstractIT {
     case Columns.ORDER_FULFILLED:
     case Columns.ANALYSIS_SKIPPED:
       return booleanColumnComparator;
+    case Columns.ROWS:
+    case Columns.COLUMNS:
+      return numericComparator;
     case Columns.LIBRARY_NAME:
     case Columns.NAME:
     case Columns.SAMPLE_NAME: