+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/rest/SamplePurposeRestController.java
@@ -35,7 +35,6 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -106,10 +105,11 @@ public class SamplePurposeRestController extends RestController {
     return getSamplePurpose(id, uriBuilder, response);
   }
 
-  @DeleteMapping(value = "/{id}")
-  @ResponseStatus(code = HttpStatus.NO_CONTENT)
-  public void deleteSamplePurpose(@PathVariable(name = "id", required = true) long id) throws IOException {
-    RestUtils.delete("Sample purpose", id, samplePurposeService);
+  @PostMapping(value = "/bulk-delete")
+  @ResponseBody
+  @ResponseStatus(HttpStatus.NO_CONTENT)
+  public void bulkDelete(@RequestBody(required = true) List<Long> ids) throws IOException {
+    RestUtils.bulkDelete("Sample Purpose", ids, samplePurposeService);
     menuController.refreshConstants();
   }