+++ miso-web/src/main/java/uk/ac/bbsrc/tgac/miso/webapp/controller/rest/StudyRestController.java
@@ -9,7 +9,6 @@ import javax.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -67,10 +66,11 @@ public class StudyRestController extends RestController {
     return RestUtils.getObject("Study", studyId, studyService, Dtos::asDto);
   }
 
-  @DeleteMapping(value = "/{id}", produces = "application/json")
-  @ResponseStatus(code = HttpStatus.NO_CONTENT)
-  public void delete(@PathVariable("id") Long id) throws IOException {
-    RestUtils.delete("Study", id, studyService);
+  @PostMapping(value = "/bulk-delete")
+  @ResponseBody
+  @ResponseStatus(HttpStatus.NO_CONTENT)
+  public void bulkDelete(@RequestBody(required = true) List<Long> ids) throws IOException {
+    RestUtils.bulkDelete("Study", ids, studyService);
   }
 
   @GetMapping(produces = "application/json")