+++ miso-web/src/main/webapp/scripts/hot_institute.js
@@ -21,26 +21,12 @@ HotTarget.institute = {
     return !config.isAdmin ? [] : [{
       name: 'Edit',
       action: function(items) {
-        window.location = window.location.origin + '/miso/institute/bulk/edit?' + jQuery.param({
+        window.location = Urls.ui.institutes.bulkEdit + '?' + jQuery.param({
           ids: items.map(Utils.array.getId).join(',')
         });
       }
     },
 
-    {
-      name: 'Delete',
-      action: function(items) {
-        var deleteNext = function(index) {
-          if (index == items.length) {
-            window.location = window.location.origin + '/miso/institute/list';
-            return;
-          }
-          Utils.ajaxWithDialog('Deleting ' + items[index].alias, 'DELETE', '/miso/rest/institutes/' + items[index].id, null, function() {
-            deleteNext(index + 1);
-          });
-        };
-        deleteNext(0);
-      }
-    }, ];
+    ListUtils.createBulkDeleteAction('Institutes', 'institutes', Utils.array.getAlias)];
   }
 };