+++ miso-web/src/main/webapp/scripts/hot_tissuematerial.js
@@ -18,31 +18,15 @@ HotTarget.tissuematerial = {
   },
 
   getBulkActions: function(config) {
-    return !config.isAdmin ? [] : [
-        {
-          name: 'Edit',
-          action: function(items) {
-            window.location = window.location.origin + '/miso/tissuematerial/bulk/edit?' + jQuery.param({
-              ids: items.map(Utils.array.getId).join(',')
-            });
-          }
-        },
+    return !config.isAdmin ? [] : [{
+      name: 'Edit',
+      action: function(items) {
+        window.location = window.location.origin + Urls.ui.tissueMaterials.bulkEdit + '?' + jQuery.param({
+          ids: items.map(Utils.array.getId).join(',')
+        });
+      }
+    },
 
-        {
-          name: 'Delete',
-          action: function(items) {
-            var deleteNext = function(index) {
-              if (index == items.length) {
-                window.location = window.location.origin + '/miso/tissuematerial/list';
-                return;
-              }
-              Utils.ajaxWithDialog('Deleting ' + items[index].alias, 'DELETE', '/miso/rest/tissuematerials/' + items[index].id, null,
-                  function() {
-                    deleteNext(index + 1);
-                  });
-            };
-            deleteNext(0);
-          }
-        }, ];
+    ListUtils.createBulkDeleteAction("Tissue Materials", "tissuematerials", Utils.array.getAlias)];
   }
 };