+++ miso-web/src/main/webapp/scripts/hot_tissueorigin.js
@@ -21,32 +21,15 @@ HotTarget.tissueorigin = {
   },
 
   getBulkActions: function(config) {
-    return !config.isAdmin ? [] : [
-        {
-          name: 'Edit',
-          action: function(items) {
-            window.location = window.location.origin + '/miso/tissueorigin/bulk/edit?' + jQuery.param({
-              ids: items.map(Utils.array.getId).join(',')
-            });
-          }
-        },
+    return !config.isAdmin ? [] : [{
+      name: 'Edit',
+      action: function(items) {
+        window.location = window.location.origin + Urls.ui.tissueOrigins.bulkEdit + '?' + jQuery.param({
+          ids: items.map(Utils.array.getId).join(',')
+        });
+      }
+    },
 
-        {
-          name: 'Delete',
-          action: function(items) {
-            var deleteNext = function(index) {
-              if (index >= items.length) {
-                window.location = window.location.origin + '/miso/tissueorigin/list';
-                return;
-              }
-              var next = function() {
-                deleteNext(index + 1);
-              };
-              Utils.ajaxWithDialog('Deleting ' + items[index].alias, 'DELETE', '/miso/rest/tissueorigins/' + items[index].id, null, next,
-                  next);
-            };
-            deleteNext(0);
-          }
-        }, ];
+    ListUtils.createBulkDeleteAction("Tissue Origins", "tissueorigins", Utils.array.getAlias)];
   }
 };