+++ miso-web/src/main/webapp/scripts/list_tissuematerial.js
@@ -31,7 +31,7 @@ ListTarget.tissuematerial = {
     return HotTarget.tissuematerial.getBulkActions(config);
   },
   createStaticActions: function(config, projectId) {
-    return config.isInternal ? [{
+    return config.isAdmin ? [{
       "name": "Add",
       "handler": function() {
 
@@ -45,7 +45,7 @@ ListTarget.tissuematerial = {
             Utils.showOkDialog('Create Tissue Materials', ["That's a peculiar number of tissuematerials to create."]);
             return;
           }
-          window.location = '/miso/tissuematerial/bulk/new?' + jQuery.param({
+          window.location = Urls.ui.tissueMaterials.bulkCreate + '?' + jQuery.param({
             quantity: result.quantity,
           });
         });