+++ RELEASE_NOTES.md
@@ -7,10 +7,12 @@ Changes:
     * Libraries: Pool Preparation, Dilution Preparation
     * Library aliquots: Pool Preparation, Dilution Preparation
   * Sort Transfers list by last modified time
+  * Allow only administrators to add tissue materials (detailed sample)
   * Added another version of the Index Distance tool to the MISO online documentation site
   * Fixed an issue where Run Scanner could cause a container to repeatedly flip between multiple
     container models if the container is linked to multiple runs
   * Fixed error when creating Sample Classes
+  * Fixed several bugs related to deleting things
 
 # 1.0.1